+++ Static.cpp
@@ -4,7 +4,7 @@
 Static::Static(int rows, int cols, Neopixel* strip) : Animation(rows, cols, strip){
 }
 
-void Static::animate(float t, float dt){
+void Static::animate(long t, float dt){
     Neopixel::IntRGB rgb = Neopixel::convertToInt(Neopixel::hsv2rgb({param0 * 360.0f, param1, 1.0f}));
     int column, row;
     for(column = 0; column < columns; column++){
